[Refactor] Create a pipeline object for applying visual effects
Due to the fact that the bulk of the next changes will be a series of sequential steps applied to the mask
the post processing function has been moved to a separate class. On top of that a protocol for applying effects was added, which nicely splits the code in to self-explanatory files.